+++ packages/@guardian/source-react-components/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @guardian/source-react-components
 
+## 7.1.3
+
+### Patch Changes
+
+- fee8fd5b: the width of the button loading spinner icon is set to 24px as it was previously too large
+
+## 7.1.2
+
+### Patch Changes
+
+- 9d7fa178: - #1548 Fix missing label for `newsletter` icon
+
 ## 7.1.1
 
 ### Patch Changes